+++ examples/init.php
@@ -16,6 +16,23 @@ $connectionOptions = [
 
     // normal unencrypted connection via TCP/IP
     ConnectionOptions::OPTION_ENDPOINT => 'tcp://localhost:8529',  // endpoint to connect to
+    
+    // // to use failover (requires ArangoDB 3.3 and the database running in active/passive failover mode)
+    // // it is possible to specify an array of endpoints as follows:
+    // ConnectionOptions::OPTION_ENDPOINT    => [ 'tcp://localhost:8531', 'tcp://localhost:8532' ]
+    
+    // // to use memcached for caching the currently active leader (to spare a few connection attempts
+    // // to followers), it is possible to install the Memcached module for PHP and set the following options:
+    // // memcached persistent id (will be passed to Memcached::__construct)
+    // ConnectionOptions::OPTION_MEMCACHED_PERSISTENT_ID => 'arangodb-php-pool',
+    // // memcached servers to connect to (will be passed to Memcached::addServers)
+    // ConnectionOptions::OPTION_MEMCACHED_SERVERS       => [ [ '127.0.0.1', 11211 ] ],
+    // // memcached options (will be passed to Memcached::setOptions)
+    // ConnectionOptions::OPTION_MEMCACHED_OPTIONS       => [ ],
+    // // key to store the current endpoints array under
+    // ConnectionOptions::OPTION_MEMCACHED_ENDPOINTS_KEY => 'arangodb-php-endpoints'
+    // // time-to-live for the endpoints array stored in memcached
+    // ConnectionOptions::OPTION_MEMCACHED_TTL           => 600
 
     // // connection via SSL
     // ConnectionOptions::OPTION_ENDPOINT        => 'ssl://localhost:8529',  // SSL endpoint to connect to